git-svn-id: http://javascool.googlecode.com/svn@1252 17e3714e-118d-3e7a-01ef-bee73ecf6174

+++ dev5/bin/sketchbook4to5.sh
@@ -22,22 +22,25 @@ xslt="`pwd`/work/jsrc/builder/hdoc2htm.xslt"
 # Boucle sur les proglets
 
 for p in $srcProglets
-do s="$srcDir/$p" ; d="$dstDir/$p"
+do s="$srcDir/$p" ; d="$dstDir/$p" ; echo "Translate $p"
 
   # Copie en miroir des fichiers
   /bin/rm -rf $d ;  svn -q export --force $s $d
   
   # Translation des fichiers de ressource
-  pushd $d
+  pushd $d > /dev/null
   if [ -f proglet.pml ] ; then java -cp $jvs org.javascool.tools.Pml proglet.pml proglet.json ; rm proglet.pml ; else echo "Erreur : y'a pas de proglet.pml !" ; fi
   if [ -f completion.xml ] ; then java -cp $jvs org.javascool.tools.Pml completion.xml completion.json ; rm completion.xml ; fi
   # Translation des fichiers *.html
-  for f in *.xml ; do n=`echo $f | sed s'/.xml$//'` ; java -jar $jsx -o $n.html $n.xml $xslt ; rm $f ; done
+  if ls *.xml > /dev/null ; then for f in *.xml ; do n=`echo $f | sed s'/.xml$//'` ; java -jar $jsx -o $n.html $n.xml $xslt ; rm -f $f ; done ; fi
     for f in *.html ; do cp $f $f~ ; sed s'/.htm"/.html"/g' < $f~ > $f ; done
-    rm -f *~
+  # Manipulation des fichiers java
+  if ls *.java > /dev/null ; then for f in *.java ; do cp $f $f~ ; cat $f~ |\
+     sed 's/org\.javascool/org.javascool/g' > $f
+  done ; fi
+  rm -f *~
   popd > /dev/null
 
-  # Listing du resultat
-  ls -l $d
-
 done
+
+svn status .
